Finally got rid of the stupid 'TruePlayer' nonsense
